third_party/blink/web_tests/platform/generic/accessibility/chromium-only-roles-expected.txt

This test makes sure that a generic focusable div can get accessibility focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axContainer.childAtIndex(0).role is "AXRole: AXSplitter"
PASS axContainer.childAtIndex(1).role is "AXRole: AXParagraph"
PASS axContainer.childAtIndex(2).role is "AXRole: AXLabelText"
PASS axContainer.childAtIndex(3).role is "AXRole: AXForm"
PASS axContainer.childAtIndex(4).role is "AXRole: AXGenericContainer"
PASS successfullyParsed is true

TEST COMPLETE
